+++ app/services/jogoService.py
@@ -6,20 +6,26 @@ class JogoService:
         return self.jogo_repository.listar_jogos()
 
     def buscar_jogo_por_nome(self, nome):
-        return self.jogo_repository.buscar_jogo_por_nome(nome)
+        jogo = self.jogo_repository.buscar_jogo_por_nome(nome)
+        if not jogo:
+            raise ValueError("Jogo não encontrado.")
+        else:
+            print(jogo)
+            return jogo
 
-    def adicionar_jogo(self, nome, descricao, data_lancamento, generos, empresa, preco):
-        return self.jogo_repository.adicionar_jogo(nome, descricao, data_lancamento, generos, empresa, preco)
+    def adicionar_jogo(self, nome, descricao, data, generos, empresa, preco):
+        if self.jogo_repository.buscar_jogo_por_nome(nome):
+            raise ValueError("Já existe um jogo com esse nome.")
+        return self.jogo_repository.adicionar_jogo(nome, descricao, data, generos, empresa, preco)
 
-    def adicionar_genero_ao_jogo(self, jogo, genero):
-        self.jogo_repository.adicionar_genero_ao_jogo(jogo, genero)
+    def atualizar_jogo(self, nome_antigo, novo_nome, descricao, data, generos, empresa, preco):
+        jogo = self.jogo_repository.buscar_jogo_por_nome(nome_antigo)
+        if not jogo:
+            raise ValueError("Jogo a ser atualizado não foi encontrado.")
+        return self.jogo_repository.atualizar_jogo(nome_antigo, novo_nome, descricao, data, generos, empresa, preco)
 
-    def listar_generos_do_jogo(self, jogo):
-        return self.jogo_repository.listar_generos_do_jogo(jogo)
-
-    def atualizar_jogo(self, nomeJogo, novo_nome, nova_descricao, nova_data, novos_generos, nova_empresa, novo_preco):
-        return self.jogo_repository.atualizar_jogo(nomeJogo, novo_nome, nova_descricao, nova_data, novos_generos, nova_empresa, novo_preco)
-    
     def remover_jogo(self, nome):
+        jogo = self.jogo_repository.buscar_jogo_por_nome(nome)
+        if not jogo:
+            raise ValueError("Jogo não encontrado.")
         return self.jogo_repository.remover_jogo(nome)
-
